feat: Rename new network config parameter to 'milestoneInterval' + Remove unused code (state) + Refactor cleanup
