+++ api/src/routes/networks/get.ts
@@ -32,7 +32,7 @@ export async function get(config: IConfiguration): Promise<INetworkGetResponse> 
             bechHrp: n.bechHrp,
             didExample: n.didExample,
             faucet: n.faucet,
-            milestoneFrequencyTarget: n.milestoneFrequencyTarget
+            milestoneInterval: n.milestoneInterval
         }))
     };
 }
