+++ client/src/app/routes/stardust/LandingState.ts
@@ -80,9 +80,4 @@ export interface LandingState extends CurrencyState, FeedsState {
      * Blocks snapshot when feed is paused.
      */
     frozenBlocks: IFeedItem[];
-
-    /**
-     * The number of seconds passed since last milestone.
-     */
-    secondsSinceLastMilestone?: number;
 }
