+++ client/src/models/config/INetwork.ts
@@ -116,7 +116,7 @@ export interface INetwork {
     faucet?: string;
 
     /**
-     * Targeted time in seconds to add latest milestone.
+     * Targeted interval in seconds between milestones.
      */
-    milestoneFrequencyTarget?: number;
+    milestoneInterval?: number;
 }
